requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
astroid==1.6.4
autopep8==1.3.5
backports.functools-lru-cache==1.5
billiard==3.5.0.4
boto3==1.4.4
botocore==1.5.95
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
cryptography==2.2.2
Django==1.11.15
django-cors-headers==1.3.1
django-debug-toolbar==1.9.1
django-extensions==1.7.4
django-filter==1.1.0
django-model-utils==3.1.2
django-safedelete==0.5.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
drf-rw-serializers==0.1.0
drfdocs==0.0.11
enum34==1.1.6
et-xmlfile==1.0.1
futures==3.2.0
hashids==1.1.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
ipaddress==1.0.22
isort==4.3.4
jdcal==1.4
jmespath==0.9.3
kombu==4.2.1
lazy-object-proxy==1.3.1
lml==0.0.1
mccabe==0.6.1
MySQL-python==1.2.5
mysqlclient==1.3.9
nexmo==1.4.0
openpyxl==2.5.3
pep8==1.7.1
phonenumbers==8.8.6
Pillow==5.2.0
pycodestyle==2.4.0
pycparser==2.18
pyexcel==0.4.2
pyexcel-io==0.5.7
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
PyJWT==1.6.4
pylint==1.9.1
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
reportlab==3.5.4
requests==2.18.4
s3transfer==0.1.13
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.2.4
texttable==1.2.1
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
wrapt==1.10.11
xhtml2pdf==0.2.2
xlrd==1.1.0
xlwt==1.3.0
